ccproject/requirements.txt
pip
Flask
cassandra-driver  --install-option="--no-cython"
passlib
flask-login
requests
